Move search function into search hooks
