+++ src/documentation/search/search-hooks.tsx
@@ -1,18 +1,34 @@
 /**
- * (c) 2021, Micro:bit Educational Foundation and contributors
+ * (c) 2022, Micro:bit Educational Foundation and contributors
  *
  * SPDX-License-Identifier: MIT
  */
-import { createContext, ReactNode, useContext, useEffect } from "react";
+import { DebouncedFunc } from "lodash";
+import debounce from "lodash.debounce";
+import {
+  createContext,
+  ReactNode,
+  useContext,
+  useEffect,
+  useMemo,
+  useState,
+} from "react";
+import useIsUnmounted from "../../common/use-is-unmounted";
+import { useLogging } from "../../logging/logging-hooks";
 import { useToolkitState } from "../toolkit-hooks";
-import { Search } from "./common";
+import { Search, SearchResults } from "./common";
 import { WorkerSearch } from "./search-client";
 
 const search: Search = new WorkerSearch();
 
-const SearchContext = createContext<Search | undefined>(undefined);
+type UseSearch = [
+  SearchResults | undefined,
+  DebouncedFunc<(newQuery: string) => Promise<void>>
+];
 
-export const useSearch = (): Search => {
+const SearchContext = createContext<UseSearch | undefined>(undefined);
+
+export const useSearch = (): UseSearch => {
   const value = useContext(SearchContext);
   if (!value) {
     throw new Error("Missing provider!");
@@ -22,6 +38,9 @@ export const useSearch = (): Search => {
 
 const SearchProvider = ({ children }: { children: ReactNode }) => {
   const { exploreToolkit, referenceToolkit } = useToolkitState();
+  const [results, setResults] = useState<SearchResults | undefined>();
+  const isUnmounted = useIsUnmounted();
+  const logging = useLogging();
   useEffect(() => {
     // Wait for both, no reason to index with just one then redo with both.
     if (exploreToolkit.status === "ok" && referenceToolkit) {
@@ -29,8 +48,30 @@ const SearchProvider = ({ children }: { children: ReactNode }) => {
     }
   }, [exploreToolkit, referenceToolkit]);
 
+  const debouncedSearch = useMemo(
+    () =>
+      debounce(async (newQuery: string) => {
+        const trimmedQuery = newQuery.trim();
+        if (trimmedQuery) {
+          const results = await search.search(trimmedQuery);
+          if (!isUnmounted()) {
+            setResults((prevResults) => {
+              if (!prevResults) {
+                logging.event({ type: "search" });
+              }
+              return results;
+            });
+          }
+        } else {
+          setResults(undefined);
+        }
+      }, 300),
+    [setResults, isUnmounted, logging]
+  );
+
+  const value: UseSearch = [results, debouncedSearch];
   return (
-    <SearchContext.Provider value={search}>{children}</SearchContext.Provider>
+    <SearchContext.Provider value={value}>{children}</SearchContext.Provider>
   );
 };
